requirements.txt
black
sqlalchemy
PyQt6
pyqt6-tools
python-dateutil
